requirements.txt
# Skill Client System - Requirements

# Core dependencies
openai>=1.0.0              # OpenAI API client (works with compatible APIs)
python-dotenv>=1.0.0       # Environment variable management
pyyaml>=6.0.0              # YAML parsing for skill definitions

# Backend server dependencies
fastapi>=0.104.0           # Web framework for backend API
uvicorn>=0.24.0            # ASGI server for FastAPI
websockets>=12.0           # WebSocket support

# Optional dependencies for different providers
# anthropic>=0.18.0        # Anthropic Claude API (if using Claude)
# cohere>=4.0.0            # Cohere API (if using Cohere)
